+++ packages/react-server/src/ReactServerStreamConfigBrowser.js
