+++ packages/react-server/src/ReactServerStreamConfigNode.js
