+++ packages/react-server/src/forks/ReactServerStreamConfig.dom.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from '../ReactServerHostConfigNode';
+export * from '../ReactServerStreamConfigNode';
