+++ scripts/rollup/forks.js
@@ -298,7 +298,7 @@ const forks = Object.freeze({
     );
   },
 
-  'react-server/src/ReactServerHostConfig': (
+  'react-server/src/ReactServerStreamConfig': (
     bundleType,
     entry,
     dependencies,
@@ -316,11 +316,11 @@ const forks = Object.freeze({
         if (!rendererInfo.isServerSupported) {
           return null;
         }
-        return `react-server/src/forks/ReactServerHostConfig.${rendererInfo.shortName}.js`;
+        return `react-server/src/forks/ReactServerStreamConfig.${rendererInfo.shortName}.js`;
       }
     }
     throw new Error(
-      'Expected ReactServerHostConfig to always be replaced with a shim, but ' +
+      'Expected ReactServerStreamConfig to always be replaced with a shim, but ' +
         `found no mention of "${entry}" entry point in ./scripts/shared/inlinedHostConfigs.js. ` +
         'Did you mean to add it there to associate it with a specific renderer?'
     );
